+++ core-base-fe/src/index.ts
@@ -1,3 +1,13 @@
+/**
+ * This file includes components used by other CORE modules such as enhanced React Admin Fields and Inputs.
+ *
+ * @remarks
+ * The `core-base` defines the {@link DateField} interface and {@link TimeAgoField} components.
+ * which are used to build widgets.
+ *
+ * @packageDocumentation
+ */
+
 import { DateField } from './react-admin/fields/DateField'
 import { TimeAgoField, timeAgo } from './react-admin/fields/TimeAgoField'
 import { PolymorphicReferenceField } from './react-admin/fields/PolymorphicReferenceField'
@@ -6,11 +16,21 @@ import { StatusInput } from './react-admin/inputs/StatusInput'
 import { MoneyInput } from './react-admin/inputs/MoneyInput'
 import { generateTimeZoneOptions } from './react-admin/fields/TimezoneSelect'
 
-// Utilities
+
+/**
+ * function utilities
+ * @public
+ */
 export { timeAgo, generateTimeZoneOptions }
 
-// Fields
+/**
+ * React Admin Fields
+ * @public
+ */
 export { DateField, TimeAgoField, PolymorphicReferenceField, MoneyField }
 
-// Inputs
+/**
+ * React Admin Inputs
+ * @public
+ */
 export { StatusInput, MoneyInput }