+++ core-comments-fe/vite.config.ts
@@ -1,4 +1,4 @@
-import dts from '@moonlight-labs/vite-plugin-dts'
+import dts from 'vite-plugin-dts'
 import react from '@vitejs/plugin-react'
 import { resolve } from 'path'
 import autoExternal from 'rollup-plugin-auto-external'
@@ -7,7 +7,16 @@ import { defineConfig } from 'vite'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [dts({ rollupTypes: true }), react(), visualizer({ open: false })],
+  plugins: [dts({
+    rollupTypes: true,
+    rollupConfig: {
+      docModel: {
+        enabled: true,
+        apiJsonFilePath: '../docs/meta/<unscopedPackageName>.api.json'
+      }
+    }}),
+    react(),
+    visualizer({ open: false })],
   build: {
     sourcemap: true,
     lib: {