+++ core-accounting-fe/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @moonlight-labs/core-accounting-fe
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @moonlight-labs/core-base-fe@0.2.3
+
+## 0.2.2
+
+### Patch Changes
+
+- 0.2.2 bump
+- Updated dependencies
+  - @moonlight-labs/core-base-fe@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes