+++ test-core-fe/CHANGELOG.md
@@ -1,5 +1,36 @@
 # test-core-fe
 
+## 0.2.4
+
+### Patch Changes
+
+- Simplify package.exports
+- Updated dependencies
+  - @moonlight-labs/core-comments-fe@0.2.4
+  - @moonlight-labs/core-versions-fe@0.2.3
+  - @moonlight-labs/core-webhooks-fe@0.2.3
+  - @moonlight-labs/core-base-fe@0.2.3
+  - @moonlight-labs/core-jobs-fe@0.2.3
+
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @moonlight-labs/core-comments-fe@0.2.3
+
+## 0.2.2
+
+### Patch Changes
+
+- 0.2.2 bump
+- Updated dependencies
+  - @moonlight-labs/core-comments-fe@0.2.2
+  - @moonlight-labs/core-versions-fe@0.2.2
+  - @moonlight-labs/core-webhooks-fe@0.2.2
+  - @moonlight-labs/core-base-fe@0.2.2
+  - @moonlight-labs/core-jobs-fe@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes